Rename to sls_sdk and alias under serverless_sdk
