+++ python/packages/sdk/sls_sdk/lib/emitter.py
