+++ python/packages/sdk/sls_sdk/lib/error_captured_event.py
