+++ python/packages/sdk/sls_sdk/lib/notice.py
