+++ python/packages/sdk/sls_sdk/lib/stack_trace_string.py
