+++ python/packages/sdk/sls_sdk/lib/warning_captured_event.py
