+++ python/packages/sdk/tests/lib/test_error.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 import pytest
 from unittest.mock import MagicMock, patch, ANY
-from serverless_sdk.lib.error import report as report_error, logger
-import serverless_sdk.lib.error_captured_event
+from sls_sdk.lib.error import report as report_error, logger
+import sls_sdk.lib.error_captured_event
 
 
 def test_error_with_exception(monkeypatch):
@@ -10,7 +10,7 @@ def test_error_with_exception(monkeypatch):
     error = Exception("Something went wrong")
     create_error_captured_event = MagicMock()
     monkeypatch.setattr(
-        serverless_sdk.lib.error_captured_event,
+        sls_sdk.lib.error_captured_event,
         "create",
         create_error_captured_event,
     )
@@ -46,7 +46,7 @@ def test_error_with_custom_object(monkeypatch):
     }
     create_error_captured_event = MagicMock()
     monkeypatch.setattr(
-        serverless_sdk.lib.error_captured_event,
+        sls_sdk.lib.error_captured_event,
         "create",
         create_error_captured_event,
     )
